Fixed my errors in editing stage directions
In my previous edits, I changed the position of the <hi> in the stage directions from around the names of the characters to the actual stage directions because in the original version of the text, the stage directions themselves are italicized and the names are normal. For example, I changed "<stage>[She goes out, and <hi>Will.</hi> follows her.</stage>" to "<stage><hi>She goes out, and </hi> Will. <hi>follows her</hi></stage>." 
However, I forgot that the <stage> code automatically italicizes the phrase and the <hi> is to make the names NOT italicized, which is exactly how the original text was written. So, I basically just reverted the stage directions that I changed back to their original state because I got confused by the code and changed a bunch of highlights that didn't need to be changed!